extra: add DSEE-HX and increase audio quality to 32/96khz
Added DSEE and alot of other Sony audio mods. NOTE: If you exit and re-enter the app says DSEE is disabled, it is NOT! This is just a visual bug due to Sony's DRM. DSEE is working in the background! (Taken from: https://forum.xda-developers.com/t/port-all-devices-all-roms-sony-xperia-xz2-sound-system-tm.3772758/)

Increased audio quality to 32/96khz by default (This may cause issues in calls but we'll try to fix this and if we can't we will remove it and stick to the regular bit rate. (Taken from: https://forum.xda-developers.com/t/xnext-tm-walkman-v1-6-7-2-update.4192117/)

Also for now the Audio Enhancements menu can only be accessed with Activity Launcher, we'll make some entries in the settings that will start activities in the app.